==> code/5.NaturalDataFit_Perenjory.R <==
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#---- 0. METADATE: Create meta data file to Summarise our data----
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# 0.1. CREATE TEMPLATES
create_spice() # creates template from directory

# 0.2. EDIT TEMPLATES
#attributes.csv: This is where most of the user data entry will take place. 
# For each variable, its name, units, and a written description are filled in.
edit_attributes()

# access.csv: Includes a row for each file that was read in, 
# and documents the name of each file and its format.
edit_access()

# creators.csv: One row for each creator, and gives their affiliation,
# contact email, ORCID, etc
edit_creators() #

#biblio.csv: Citation information about the project, 
# as much or as little data as possible can be included, 
# but if things like bounding box coordinates are not included, 
# then when the website is generated there will not be a bounding box map generated
edit_biblio()


#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#---- 1. SET UP: Import data, create df with competiton and seed distributions----
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
rm(list = ls())

#---- 1.1. Import packages ----
#install.packages("rstan", repos = "https://cloud.r-project.org/", dependencies = TRUE)
library(rstan)
#install.packages("HDInterval")
library("HDInterval")
#install.packages("tidyverse")
library("tidyverse")
#install.packages("dplyr")
library(dplyr)
library(ggpubr)
library(ggplot2)
#rstan_options(auto_write = TRUE)
library(tidyr) #fill is part of tidyr
library(lme4)
library(car)
library(loo)
library(wesanderson) # for color palette
library(ggthemes) 
library(grid)
#---- 1.2. Import the competitive data ----

competition <- read.csv("/Users/lisabuche/Documents/Projects/Perenjori/data/Sevenello2022_competition.csv",
                        header = T,stringsAsFactors = F, sep=",",
                        na.strings=c("","NA"))
species.neigh <- names(competition)
species.neigh  <- species.neigh[!species.neigh %in% c("reserve","focal","crop", "site", "plot","treatment","year")]

competition.to.keep <- competition %>%  
  dplyr::select(all_of(c(species.neigh)))%>%
  mutate_at( species.neigh, as.numeric) %>%
  colSums(na.rm=T)
length(competition.to.keep[competition.to.keep == 0]) # check if any is 0


competition <- competition %>% 
  dplyr::filter(treatment =="OP") 

# change the format of the rows to numeric 
competition[species.neigh] <- sapply(competition[species.neigh],as.numeric)

# change na values to 0
competition[is.na(competition)] <- 0

focal.levels <- levels(as.factor(competition$focal))

#---- 1.2. Make Teasorus ----
perenjory_tesaorus <- read.csv("/Users/lisabuche/Documents/Projects/Perenjori/data/perenjory_tesaorus.csv",
                        header = T,stringsAsFactors = F, sep=",",
                        na.strings=c("","NA"))

competition.long.neigh <- competition %>%
  gather(any_of(perenjory_tesaorus$code), key="code",value="abundance") %>%
  left_join(perenjory_tesaorus) 

competition.long.focal <- competition.long.neigh %>%
  filter(focal ==code) %>%
  rename("conspecific"="abundance") %>%
  dplyr::select(-code) 

competition.long <- competition.long.neigh %>%
  aggregate(abundance ~  family + reserve + year + site + focal+ crop + site + plot + treatment, sum )%>%
  spread(family, abundance)  %>%# change this ti have grass /forb or native/exotic
  right_join(competition.long.focal)

# Only run if not family - all intra and inter together
competition.long <- competition.long.neigh %>%
  mutate(neigh= case_when(final.code=="LARO" ~ "LARO",
                          T~"Inter")) %>%
  aggregate(abundance ~  neigh + reserve + year + site + focal+ crop + site + plot + treatment, sum )%>%
  spread(neigh, abundance)  %>%# change this ti have grass /forb or native/exotic
  right_join(competition.long.focal)
head(competition.long)
#---- 1.4. Import the seeds data ----
fecundity.df <- read.csv("/Users/lisabuche/Documents/Projects/Perenjori/data/Sevenello2022_seeds.csv",
                         header = T,stringsAsFactors = F, sep=",",
                         na.strings=c("","NA"))
head(fecundity.df)
fecundity.df <- fecundity.df %>% 
  dplyr::filter(treatment =="OP")

fecundity.df$reserve <- as.factor(fecundity.df$reserve)

seeds.model <- glm(round(seeds) ~ reserve + focal  ,fecundity.df, 
                     family="poisson")

summary(seeds.model) # random Side effect not signidicant - not considered further
  
seeds.graph <- ggplot(fecundity.df,aes(x=seeds,color=focal)) + 
    geom_density() +
  xlim(0,200)

# join both compeetiton and seed 
competition.seeds <- left_join(competition.long,fecundity.df,
                               by=c("site", "focal", "crop", "reserve", "plot", "treatment"))


intrinsic.fecun <- read.csv("data/Aus.intrinsic.fecun.csv",
                            header = T,stringsAsFactors = F, sep=",",
                            na.strings=c("","NA"))

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#---- 2. Run the model for each focal----
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
run.diagnostic =1
run.stan = 1
function.grouping = 1
grouping ="group.all.hetero" #"group.all.hetero"  or "family"
scale.factor <- 1
for(Code.focal in "LARO"){ #focal.levels
 for (function.int in c(1:4)){
  print(paste(Code.focal,", function",function.int))
    function.vec <- c(0,0,0,0)
  function.vec[function.int] <- 1
  
 # data for the focal
  SpDataFocal <- competition.seeds[which(competition.seeds$focal==Code.focal),] 
  names(SpDataFocal) <- tolower(names(SpDataFocal))
  SpDataFocal <-  SpDataFocal[!is.na(SpDataFocal$seeds),]
  SpDataFocal[is.na(SpDataFocal)] <- 0
  # Next continue to extract the data needed to run the model. 
  N <- as.integer(nrow(SpDataFocal))
  Fecundity <- as.integer(SpDataFocal$seeds)  
  specific.focal <- "asteraceae"
  #---- 2. ABUDANCE MATRIX----
  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  #---- 2.1. Interaction (direct) matrix of plant with COMP ----
  
  # Now calculate the total number of plant species to use for the model, discounting
  #       any species columns with 0 abundance. Save a vector of the species names
  #       corresponding to each column for easy matching later.
  AllSpNames <- names(SpDataFocal)[!names(SpDataFocal) %in% c("focal","year","day","month","transect","plot","reserve",
                                                              "block","unique.plot","site","treatment",
                                                              "crop","flower","seeds","individual", "species","origine","functional.group",      
                                                              "family","flowercolor","final.code","flowers",                
                                                              "Seeds","laro")]
  
  if(grouping=="family"){                                                    
AllSpAbunds <- SpDataFocal %>% 
  dplyr::select(all_of(c(AllSpNames)))%>%
  mutate_at( AllSpNames, as.numeric)

SpTotals <- colSums(AllSpAbunds)
SpToKeep <- SpTotals > 30 #before for family 
sum(SpTotals)
NamesSpToKeep <- names(SpTotals[SpToKeep])
Stotal <- sum(SpToKeep)
S <- Stotal
SpMatrix <- NULL
SpMatrix <- as_tibble(matrix(NA, nrow = N, ncol = S+1)) # +1
i <- 1
SpMatrix[,S+1] <- 0
for(s in 1:length(AllSpNames)){
  if(SpToKeep[s] == 1){
    if(SpTotals[s] < 5){
      SpMatrix[,S+1] <- (SpMatrix[,S+1] + AllSpAbunds[,s])
      
    }else{
      SpMatrix[,i] <- AllSpAbunds[,s]
      i <- i + 1}
  }else{next}
}
SpMatrix <-SpMatrix/scale.factor
SpNames <- c(AllSpNames[SpToKeep],"rare")
names(SpMatrix) <-  SpNames
Intra <- ifelse(SpNames == "conspecific", 1, 0)
SpMatrix[,which(SpNames ==specific.focal)] <- SpMatrix[,which(SpNames == specific.focal)] - SpMatrix[,"conspecific"]  # remove conspecific obs from focal family

SpMatrix <- as.matrix(SpMatrix)

}else{
  AllSpAbunds <- SpDataFocal %>% 
    dplyr::select(all_of(c(AllSpNames)))%>%
    mutate_at( AllSpNames, as.numeric)
  
  SpTotals <- colSums(AllSpAbunds)
  SpToKeep <- SpTotals > 0 #before for family 
  sum(SpTotals)
  NamesSpToKeep <- names(SpTotals[SpToKeep])
  Stotal <- sum(SpToKeep)
  S <- Stotal
  SpMatrix <- NULL
  SpMatrix <- as_tibble(matrix(NA, nrow = N, ncol = S)) # +1
  i <- 1
  for(s in 1:length(AllSpNames)){
    if(SpToKeep[s] == 1){
      SpMatrix[,i] <- AllSpAbunds[,s]
      i <- i + 1
    }else{next}
  }
  SpMatrix <-SpMatrix/scale.factor
  SpNames <- c(AllSpNames[SpToKeep])
  names(SpMatrix) <-  SpNames
  Intra <- ifelse(SpNames == "conspecific", 1, 0)
  SpMatrix[,which(SpNames ==specific.focal)] <- SpMatrix[,which(SpNames == specific.focal)] - SpMatrix[,"conspecific"]  # remove conspecific obs from focal family
  
  SpMatrix <- as.matrix(SpMatrix)
  S= S-1
}
  
  Nmedian <- c(SpMatrix[which.max(Fecundity),])

  # Upper bound intrinsic fecundity
  Fmean <- intrinsic.fecun$lambda[which(intrinsic.fecun$code==Code.focal)]

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#---- 3. BAYES FIT----
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  ##---- 3.1. Set up summary interactions df and parameters ---- 
  
  run_estimation <- 1
  alphaFunct1 <- function.vec[1]
  alphaFunct2 <- function.vec[2]
  alphaFunct3 <- function.vec[3]
  alphaFunct4 <- function.vec[4]
  
  alpha.function <- paste0("function_",which(function.vec==1))
  
  if(alphaFunct1==0){
    load(paste0("results/stan/Parameters_",grouping,"_",Code.focal,"_function_1",".RData"))
    parameters <- get(paste0("Parameters_",Code.focal,"_function_1_",grouping))
   # Fmean = median(parameters$lambda)
    
  }
  DataVec <- list(N=N, 
                  S=S+1,
                  lambda_prior = Fmean,
                  Nmedian = Nmedian ,
                  Fecundity=Fecundity,
                  SpMatrix =SpMatrix,
                  Intra=Intra,
                  run_estimation=run_estimation,
                  alphaFunct1=alphaFunct1,
                  alphaFunct2=alphaFunct2,
                  alphaFunct3=alphaFunct3,
                  alphaFunct4=alphaFunct4
                  )
  
  ##---- 3.2. Run  final fit ----
  # Now run a fianl fit of the model to assess parameter 
  print("Natural Fit beginning")
  
  #install.packages("codetools")
  library("codetools")
  options(mc.cores = parallel::detectCores())
  # defiining initial value of lambda to help with computation failure
  # divide by the maximum expected for lambda = U 
  list.init <- function(...)list(lambda= array(as.numeric(Fmean), dim = 1))
  
  if( run.stan == 1){                               
  FinalFit <- stan(file = "code/DensityFunct_Final.stan", 
                   data = DataVec,
                   init=  list.init,
                   #init =  "random",
                   warmup= 1000,
                   iter = 3000, 
                   init_r = 2,
                   chains = 4,
                   control=list(max_treedepth=15),
                   seed= 16)
  
  
  save(file= paste0("results/stan/FinalFit_",grouping,"_",Code.focal,"_",alpha.function,".rds"),
       FinalFit)
  
  if(length(rstan::extract(FinalFit)) == 0){ # restrict less niital values
   FinalFit <- stan(file = "code/DensityFunct_Final.stan", 
                     data = DataVec,
                     init =  "random",
                     warmup= 1000,
                     iter = 3000, 
                     init_r = 2,
                     chains = 4,
                     control=list(max_treedepth=15),
                     seed= 16)
   save(file= paste0("results/stan/FinalFit_",grouping,"_",Code.focal,"_",alpha.function,".rds"),
        FinalFit)
    
  }
  }
  #load(paste0("results/stan/FinalFit_",Code.focal,"_",alpha.function,".rds"))
  
 load(paste0("results/stan/FinalFit_",grouping,"_",Code.focal,"_",alpha.function,".rds"))

 FinalPosteriors <- rstan::extract(FinalFit)
  
  print("Final Fit done")
  
  #---- 3.3. Final fit posterior check and behavior checks---- 
  if( run.diagnostic == 1){  
    if(function.int == 1){par <- c('lambda',
                                   'alpha_initial')
    }
    if(function.int ==2){par <- c('lambda',
                                  'alpha_initial',
                                  'alpha_slope')
    }
    if(function.int >2){par <- c('lambda',
                                 'c','alpha_initial',
                                 'alpha_slope','N_opt')
    }
    
  ##### Diagnostic plots and post prediction 
  pdf(paste0("figures/stan/FinalFit_",grouping,"_",Code.focal,"_",alpha.function,".pdf"))
  # Internal checks of the behaviour of the Bayes Modelsummary(PrelimFit)
  source("code/6.1.Stan_modelcheck_rem.R") # call the functions to check diagnistic plots
  # check the distribution of Rhats and effective sample sizes 
  ##### Posterior check
  stan_post_pred_check(FinalPosteriors,"F_hat",Fecundity,
                       paste0("results/stan/PostFec_",grouping,
                              "_",Code.focal,"_",alpha.function,".csv.gz")) 
  
  #log_lik_2 <- loo::extract_log_lik(FinalFit, 
  #                                  parameter_name = "F_sim", 
  #                                  merge_chains = F)
  
  #r_eff <- loo::relative_eff(exp(log_lik_2), cores = 2) 
  
 # loo_1 <- loo::loo(log_lik_2 , r_eff = r_eff, cores = 2)
  
 # print(loo_1)
  
  # N.B. amount by which autocorrelation within the chains increases uncertainty in estimates can be measured
  hist(summary(FinalFit)$summary[,"Rhat"],
       main = paste("Finat Fit: Histogram of Rhat for",
                    Code.focal," and ",alpha.function))
  hist(summary(FinalFit)$summary[,"n_eff"],
       main = paste("Finat Fit: Histogram of Neff for",
                    Code.focal," and ",alpha.function))
  
  # plot the corresponding graphs
  trace <- stan_trace(FinalFit, pars=par,
                      inc_warmup = TRUE)
  print(trace)
  dens <- stan_dens(FinalFit, 
                    pars=par)
  print(dens)
  splot <- stan_plot(FinalFit, 
                     pars=par)
  print(splot)
  sampler_params <- get_sampler_params(FinalFit, inc_warmup = TRUE)
  summary(do.call(rbind, sampler_params), digits = 2)
  pairs(FinalFit, pars = par)
  
  # Next check the correlation among key model parameters and identify any
  #pairs(FinalFit, pars = c("lambda",'alpha_initial','alpha_slope','c'))
  
  
  dev.off()
  }
  #---- 3.3. Extract coefficients ----
  
  assign(paste0("Parameters_",Code.focal,"_",alpha.function,"_",grouping),
       list(DataVec = DataVec,
            SpNames=SpNames,
       alpha_value= FinalPosteriors$alpha_value,      
       lambda =  FinalPosteriors$lambda,
       alpha_slope = FinalPosteriors$alpha_slope,
       alpha_init =FinalPosteriors$alpha_init,
       alpha_c = FinalPosteriors$c,
       N_opt = FinalPosteriors$N_opt
       ))
  
  save(list =paste0("Parameters_",Code.focal,"_",alpha.function,"_",grouping),
       file = paste0("results/stan/Parameters_",grouping,"_",Code.focal,"_",alpha.function,".RData"))
  
  #---- 3.3. Extraction fecundity---
  
  assign(paste0("Fsim_",Code.focal,"_",alpha.function,"_",grouping),
         list(DataVec = DataVec,
              F_sim=FinalPosteriors$F_sim
         ))
  
  save(list =paste0("Fsim_",Code.focal,"_",alpha.function,"_",grouping),
       file = paste0("results/stan/Fsim_",grouping,"_",Code.focal,"_",alpha.function,".RData"))

}
}

#---- 4. Figures ----
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

#---- 4.0. check models ----
ModelCheck <- NULL
grouping<- "family" #"family"
for(Code.focal in "LARO"){#focal.levels
    for (function.int in c(1:4)){ # c(1:4)
      print(paste(Code.focal,", function",function.int))
      
      alpha.function <- paste0("function_",function.int)
      
      load(paste0("results/stan/FinalFit_",grouping,"_",Code.focal,"_",alpha.function,".rds"))
      
      mc <- data.frame(focal =Code.focal, 
                        function.int = function.int,
                        Rhat = max(summary(FinalFit)$summary[,"Rhat"],na.rm =T),
                        Neff = min(summary(FinalFit)$summary[,"n_eff"],na.rm = T))
      ModelCheck <- bind_rows(ModelCheck,mc)
      remove(FinalFit)
    }
  }
View(ModelCheck)

write.csv(ModelCheck,
          paste0("results/ModelCheck_LARO_",grouping,".csv"))

#---- 4.1 Alpha distribution ----
grouping<-"family"#// "group.all.hetero" #"family"
df_alpha_all <- data.frame(focal=NA)
for(Code.focal in "LARO" ){ #focal.levels
  for (function.int in c(1:4)){
    df_alpha <- data.frame()
    function.vec <- c(0,0,0,0)
    function.vec[function.int] <- 1
    alpha.function <- paste0("function_",which(function.vec==1))

    load(paste0("results/stan/Parameters_",grouping,"_",Code.focal,"_",alpha.function,".RData"))
    parameters <- get(paste0("Parameters_",Code.focal,"_",alpha.function,"_",grouping))
    neighbours.vec <- c(parameters[["SpNames"]])
    
    N <- parameters[["DataVec"]]$N
    Nmax <- parameters[["DataVec"]]$Nmax
    df_alpha_init <- data.frame(observation = c(1:N),
                                 focal = Code.focal,
                                 function.int = function.int)
     df_alpha_init <- NULL
     for(n in 1:length(neighbours.vec)){
       df_alpha_init_n <- data.frame(focal = Code.focal,
                                     Nmax=parameters[["N_opt"]][,n],
                                     function.int = function.int,
                                     number = c(1:length(parameters[["alpha_init"]][,n])),
                                     neigh = neighbours.vec[n],
                                     alpha_init= parameters[["alpha_init"]][,n])
       df_alpha_init <- bind_rows(df_alpha_init,df_alpha_init_n)
        }

    if (function.int > 1){
      df_alpha_slope <- NULL
      for(n in 1:length(neighbours.vec)){
        df_alpha_slope_n <- data.frame(focal = Code.focal,
                                      neigh = neighbours.vec[n],
                                      function.int = function.int,
                                      number = c(1:length(parameters[["alpha_slope"]][,n])),
                                      alpha_slope = parameters[["alpha_slope"]][,n])
        df_alpha_slope <- bind_rows(df_alpha_slope,df_alpha_slope_n)
      }
      
     if(function.int > 2){
       df_alpha_c <- NULL
       for(n in 1:length(neighbours.vec)){
         df_alpha_c_n <- data.frame(focal = Code.focal,
                                        neigh = neighbours.vec[n],
                                        function.int = function.int,
                                        number = c(1:length(parameters[["alpha_c"]][,n])),
                                        alpha_c = parameters[["alpha_c"]][,n])
         df_alpha_c <- bind_rows(df_alpha_c,df_alpha_c_n)
       }
    
    df_alpha <- full_join(df_alpha_c,full_join(df_alpha_init,df_alpha_slope))
     }else{
    df_alpha <- full_join(df_alpha_init,df_alpha_slope)
     }
    }else{
    df_alpha <- df_alpha_init
    }
     df_alpha_all <- full_join(df_alpha_all,df_alpha)
  }
}
df_alpha_all <- df_alpha_all[-1,] # nrow = 120 == 4(fct) * 5(species) * 6(neighbours)
str(df_alpha_all)
view(df_alpha_all)
levels(as.factor(df_alpha_all$neigh))
source("code/1.1.PopProjection_toolbox.R")
neigh.levels <- levels(as.factor(df_alpha_all$neigh))
df_funct_alpha <- NULL
#for (n in 1:nrow(df_alpha_all)){

for( fct.int in 1:4){
  for (n in neigh.levels){
  #print(n)
  df_funct_alpha_n <- df_alpha_all %>% 
    dplyr::filter(neigh==n 
                  & function.int ==fct.int) %>%
    dplyr::select(alpha_init,alpha_slope,alpha_c,Nmax) %>%
    dplyr::summarise_all(list(~ median(., na.rm = TRUE)))
  df_funct_alpha_n <- data.frame(density=c(0:10),df_funct_alpha_n)
  if(fct.int==1){
    df_funct_alpha_n$alpha_value <- median(df_funct_alpha_n$alpha_init)
    
    }
  if(fct.int==2){
    df_funct_alpha_n$alpha_value <- alpha_function2(df_funct_alpha_n$alpha_init,
                                                   df_funct_alpha_n$alpha_slope,
                                                    df_funct_alpha_n$density,
                                                    df_funct_alpha_n$Nmax)
    
  }
  if(fct.int==3){
    df_funct_alpha_n$alpha_value <- alpha_function3(df_funct_alpha_n$alpha_init,
                                                    df_funct_alpha_n$alpha_slope,
                                                   df_funct_alpha_n$alpha_c,
                                                  df_funct_alpha_n$density,
                                                  df_funct_alpha_n$Nmax)
    
  }
  if(fct.int==4){
    df_funct_alpha_n$alpha_value <- alpha_function3(median(df_funct_alpha_n$alpha_init),
                                                    median(df_funct_alpha_n$alpha_slope),
                                                    median(df_funct_alpha_n$alpha_c),
                                                    df_funct_alpha_n$density,
                                                    median(df_funct_alpha_n$Nmax))
  }
  
  df_funct_alpha <- bind_rows(df_funct_alpha,df_funct_alpha_n %>%
                                mutate(neigh=n,
                                       function.int =fct.int,
                                       function.name = c("1.Traditional","2.Linear","3.Exp","4.Sigmoid")[fct.int],
                                       focal="LARO"))
  
}
}
write.csv(df_funct_alpha,
          paste0("results/df_funct_alpha_LARO_",grouping,".csv.gz"))

df_funct_alpha <- data.table::fread(paste0("results/df_funct_alpha_LARO_",grouping,".csv.gz"))

cbp2 <- c("#000000", "#E69F00", "#56B4E9","#F0E442","#009E73",
          "#CC79A7", "#0072B2", "#D55E00")
if(grouping=="group.all.hetero"){
cbp2 <- c("#000000","grey")
}
cb2_focal <- c("#F0E442","#009E73",
               "#CC79A7", "#0072B2")
funct.col <- c("#AA4499", "#DDCC77","#88CCEE", "#44AA99")
plot.list.alpha <- list()


# percentage of positive interaction with function 4
100*nrow(df_funct_alpha[which(df_funct_alpha$function.int == 4 &
                       df_funct_alpha$alpha_value > 0),])/
  nrow(df_funct_alpha[which(df_funct_alpha$function.int == 4),])


# Only LARO

dummy <- data.frame(uplimit=c(0.08,0.25,0.30,0.15),
                    downlimit=c(-0.08,-0.25,-0.1,-0.1),
                    function.name = c("1.Traditional","2.Linear","3.Exp","4.Sigmoid"),
                    label.title= c("a. Traditional constant",
                                   "b. Linear","c. Exponential",
                                   "d. Sigmoid"),
                    margindown = c(-1,-1,1,1),
                    stringsAsFactors=FALSE)

plot_LAROalpha_list <- list()
family.neigh <- levels(as.factor(df_funct_alpha$neigh))
family.neigh <- family.neigh[!family.neigh=="conspecific"]
for( i in c("1.Traditional","2.Linear","3.Exp","4.Sigmoid")){
  df <- df_funct_alpha %>%
    dplyr::filter(focal == "LARO" & function.name==i)%>%
    mutate(neigh = case_when(neigh =="conspecific" ~ "LARO",
                             T ~ neigh)) %>%
    mutate(neigh= factor(neigh, 
                            levels = c(family.neigh,"LARO")))

  plot_LAROalpha_list[[i]] <- ggplot(df,
                                       aes(x=density, y= alpha_value,
      color=neigh,fill=neigh,size=neigh,alpha=neigh)) +
    geom_hline(yintercept=0, color="black", linetype="dashed",alpha=0.6,
               size=2) +
    geom_line(stat="smooth",method = 'gam', # remove these line for function 1
              se = TRUE,level =0.95,# remove these line for function 1
              alpha=0.8,# remove these line for function 1
              size=3) +
    scale_color_manual("Neighbours identity",values=rev(cbp2)) + 
    scale_fill_manual("Neighbours identity",
                      values=rev(cbp2)) + 
    scale_size_manual("Neighbours identity",
                      values=c(3, rep(1.5,length(family.neigh)))) + 
    scale_alpha_manual("Neighbours identity",
                       values=c(1, rep(0.8,length(family.neigh)))) + 
    theme_bw() +
    labs(title = dummy$label.title[which(dummy$function.name==i)] ,
      y= "Per capita effect of neighbours on focal",
      x="density of neighbours") +
    rremove("ylab") + rremove("xlab") + 
    scale_x_continuous( expand= c(0,0),
                        minor_breaks = NULL,
                        breaks=c(0,5,10),
                        labels=c("0","5","10")) +
    scale_y_continuous(expand= c(0,0),
                       minor_breaks = NULL) +
    guides(size="none",alpha="none") +
    coord_cartesian( xlim = NULL, 
                     #ylim=c(-1,1),
                     ylim = c(dummy$downlimit[which(dummy$function.name==i)],
                     dummy$uplimit[which(dummy$function.name==i)]),
                     expand = TRUE, default = FALSE, clip = "on") +
    theme(plot.title = element_text(color="#000000",
                                    vjust = 0,
                                    size=24,face = "bold"),
          strip.text =element_text(color="black",size=20),
          legend.key= element_rect(fill = "white"),
          legend.key.size = unit(1, 'cm'),
          legend.text=element_text(size=20),
          legend.background = element_blank(),
          legend.box.background = element_rect(colour = "lightgrey"),
          legend.position = "bottom",
          legend.title=element_text(size=20),
          axis.text.x = element_text(color="black",size=16),
          axis.text.y = element_text(color="black",size=16),
          plot.margin = unit(c(0,0.5,
                               dummy$margindown[which(dummy$function.name==i)],
                               0.5), 'cm')) 
  
}
plot_LAROalpha_list[[4]]
plot_LAROalpha_list[[1]]
plot_LAROalpha <- ggarrange( plotlist = plot_LAROalpha_list,
                              ncol=2,nrow=2,common.legend = T,
                             legend = "bottom",
                             align = c("hv"))
plot_LAROalpha.2 <- annotate_figure(plot_LAROalpha , 
                                  left = textGrob("Per capita effect on LARO", 
                                                  rot = 90, vjust = 0.9, hjust=0.4, 
                                                  gp = gpar(fontsize=20,cex = 1.3)),
                                  bottom = textGrob("Density of neighbours", 
                                                    vjust=-5.2, hjust=0.4,
                                                    gp = gpar(fontsize=20,cex = 1.3))) 
plot_LAROalpha.2
ggsave(plot_LAROalpha.2,
       file = "figures/NatData_plot_LARO.pdf")

#---- 4.2. Population Projections ----

df_param_all <- data.frame(focal=NA)
grouping ="family"
for(Code.focal in "LARO"){
  for (function.int in c(1:4)){
    #if (Code.focal == "TROR") next
    df_param <- data.frame()
    function.vec <- c(0,0,0,0)
    function.vec[function.int] <- 1
    alpha.function <- paste0("function_",which(function.vec==1))
    
    load(paste0("results/stan/Parameters_",grouping,"_",Code.focal,"_",alpha.function,".RData"))
    parameters <- get(paste0("Parameters_",Code.focal,"_",alpha.function,"_",grouping))
    N <- parameters[["DataVec"]]$N
    Nmax <- parameters[["DataVec"]]$Nmax
    neighbours.vec <- c(parameters[["SpNames"]])
    df_param_init <- data.frame(observation = c(1:N),
                                focal = Code.focal,
                                function.int = function.int)
    df_param_init <- NULL
    for(n in 1:length(neighbours.vec)){
      df_param_init_n <- data.frame(focal = Code.focal,
                                    Nmax=median(parameters[["N_opt"]][,n]),
                                    function.int = function.int,
                                    neigh = neighbours.vec[n],
                                    lambda= median(parameters[["lambda"]]),
                                    alpha_init= median(parameters[["alpha_init"]][,n]))

      df_param_init <- bind_rows(df_param_init,df_param_init_n)
    }
    
    if (function.int > 1){
      df_param_slope <- NULL
      for(n in 1:length(neighbours.vec)){
        df_param_slope_n <- data.frame(focal = Code.focal,
                                       neigh = neighbours.vec[n],
                                       function.int = function.int,
                                       alpha_slope = median(parameters[["alpha_slope"]][,n]))
        df_param_slope <- bind_rows(df_param_slope,df_param_slope_n)
      }
      
      if(function.int > 2){
        df_param_c <- NULL
        for(n in 1:length(neighbours.vec)){
          df_param_c_n <- data.frame(focal = Code.focal,
                                     neigh = neighbours.vec[n],
                                     function.int = function.int,
                                     alpha_c = median(parameters[["alpha_c"]][,n]))

          df_param_c <- bind_rows(df_param_c,df_param_c_n)
        }
        
        df_param <- full_join(df_param_c,full_join(df_param_init,df_param_slope))
      }else{
        df_param <- full_join(df_param_init,df_param_slope)
      }
    }else{
      df_param <- df_param_init
    }
    df_param_all <- full_join(df_param_all,df_param)
  }
}
df_param_all <- df_param_all[-1,] # nrow = 120 == 4(fct) * 5(species) * 6(neighbours)
str(df_param_all)
df_projection <- NULL


#----Community abundance DATA---
community_id_df <- data.table::fread("/Users/lisabuche/Documents/Projects/Perenjori/results/community_id_df_short.csv")
str(community_id_df) # already scaled by scale.width

levels(as.factor(community_id_df$family))
levels(as.factor(community_id_df$year))
scale.width.local.pred = 15 # define the width of local population 
# make df only for focal abundance over time

fam.to.keep <- levels(as.factor(df_param_all$neigh))
fam.to.keep <- fam.to.keep[which(!fam.to.keep =="rare")]

fam.to.keep <- c("araliaceae","asteraceae","conspecific",
                 "crassulaceae","goodeniaceae","montiaceae","poaceae")        
# select only family and sum per family

community_id_df_fam <- community_id_df %>% 
  mutate(groups = case_when(!family %in% fam.to.keep ~ "rare",
                           T ~ family)) %>%
  #dplyr::filter(family %in% levels(as.factor(df_param_all$neigh))) %>%
  aggregate(count ~ id.plot + collector+ year + groups + scale.width, sum) %>%
  mutate(count=count*scale.width.local.pred ) %>%
  filter(count < 1000) %>%
  spread(groups,count)
str(community_id_df_fam)

# select only LARO

community_id_df_LARO <- community_id_df %>% 
  dplyr::filter(code=="LARO") %>%
  aggregate(count ~ code + year + id.plot + collector + 
              family + scale.width, function(x) sum(x,na.rm=T)) %>%
  mutate(count=(count)*scale.width.local.pred )

#----4.2.1.function to sample in the neighbours populations----
library(MASS)  #Code for univariate poisson - to sample a initial value for LARO
abundance.sampling.poisson <- function(a,n){
  a <- a[!is.na(a)]
  lambda.neigh <- MASS::fitdistr(a,"poisson")
  ab.vec <- rpois(n,
        lambda=lambda.neigh$estimate)
  return(ab.vec)
}

# code to fund the multivariate Poisson-lognormal 
#emotes::install_github("pln-team/PLNmodels")
#install.packages("PLNmodels")
library(PLNmodels)
library(ggplot2)
library(corrplot)

test.list <- list(Abundance = community_id_df_fam %>% 
                    dplyr::select(any_of(c(fam.to.keep,"rare"))),
                  Covariate =community_id_df_fam %>% 
                    dplyr::select(any_of(c("year"))))
prepared.df <- prepare_data(counts= test.list$Abundance, 
                            covariates =  test.list$Covariate)
                             
myPLN <- PLN(Abundance ~ 1, prepared.df)
 
myPLN %>% sigma() %>% cov2cor() %>% corrplot(type="full")
#file = "figures/Natcom_corplotAbundance.pdf"
ggsave(data.frame(fitted   = as.vector(fitted(myPLN)),
  observed = as.vector(prepared.df$Abundance)) %>% 
  ggplot(aes(x = observed, y = fitted)) + 
  geom_point(size = .5, alpha =.25 ) + 
  scale_x_log10() + 
  scale_y_log10() + 
  geom_abline(slope=1,intercept=0,color="red") +
  theme_bw() + annotation_logticks() +theme(axis.text = element_text(size=15),axis.title = element_text(size=20)),
  file="figures/Natcoom_PLNfit.pdf")

#----4.2.2.Projection ----
df_projection <- NULL
gens = 20
scale.width.local.pred  =15

for(Code.focal in "LARO"){
  for(i in 1:100){
    print(i)
    # produce fake data of abundance of neighbours based on their 
    # abundance relation establish with PLN. 
    # We set asteraceae and estimated the density of others based on it
    id.plot.level <- levels(as.factor(community_id_df_fam$id.plot[which(community_id_df_fam$year==2023 | community_id_df_fam$year==2022)]))
    starting.df <- community_id_df_fam[which(community_id_df_fam$year==2023 |community_id_df_fam$year==2022),] %>%
      filter(id.plot ==  id.plot.level[sample(1:length( id.plot.level),1)]) %>%
      dplyr::select(any_of(levels(as.factor(df_param_all$neigh)))) %>%
      mutate(year=2023) %>%
      replace(is.na(.), 0)
    
    mean.a <- mean(community_id_df_fam$asteraceae,na.rm=T) 
    sd.a <-sd(community_id_df_fam$asteraceae,na.rm=T)
    state_df <- data.frame(asteraceae = round(abs(rnorm(gens,mean.a,sd.a)))) %>%
            bind_cols(round(predict_cond(myPLN,newdata =data.frame(year=c(2024:(2024+gens-1))),
                                        cond_responses = .,
                                        type = c("response"),var_par = FALSE)))
    state_df <- bind_rows(starting.df,state_df) 
      
    # matrix of 21 by 7 
    num.0 <- round(runif(1,1,gens*7/5)) # allow up to 20% of 0
    vec.0.row <- round(runif( num.0,1,gens))
    vec.0.col <- round(runif( num.0,1,7))
    for(i0 in 1:num.0){
   state_df[vec.0.row[i0] ,vec.0.col[i0]] <-0 
    }
    state_df$conspecific <- abundance.sampling.poisson(round(mean(community_id_df_LARO$count)),1)
    if(state_df$conspecific[1] ==0){
      state_df$seed.total.conspecific <- 1
    }else{
      state_df$seed.total.conspecific<- state_df$conspecific/0.2 # germination rate
    }
if(grouping=="group.all.hetero" ){
  state_df <- state_df %>%
    mutate(inter = rowSums(.[,c("araliaceae","asteraceae",
                                "crassulaceae" ,"goodeniaceae",
                                "montiaceae" ,"poaceae","rare")],na.rm=T)) %>%
    dplyr::select(-c("araliaceae","asteraceae",
                     "crassulaceae" ,"goodeniaceae",
                     "montiaceae" ,"poaceae","rare"))
    
}
for (function.int in c(1:4)){
    param.df <- df_param_all[which(df_param_all$function.int==function.int & 
                                     df_param_all$focal==Code.focal ),]
    param.df$g = 0.2
    param.df$s = 0.4
   df_projection_n <- Ricker_solution_NatData(gens=gens,
                                             state=state_df,
                                             pars = param.df,
                                             scalewidth = scale.width.local.pred ,
                                             neigh.vec = levels(as.factor(param.df$neigh)))

  df_projection_n$sim <- i 
  df_projection_n$function.int <- function.int
  df_projection_n$focal <- Code.focal
  df_projection_n$year<- c(2023:(2023+gens))
  
  df_projection <- bind_rows(df_projection, df_projection_n)
  }
  }
 }

df_projection <- df_projection %>%
  mutate(function.name = case_when(function.int==1 ~"1.Traditional",
                                   function.int==2 ~"2.Linear",
                                   function.int==3 ~"3.Exp",
                                   function.int==4 ~"4.Sigmoid")) %>%
  gather(all_of(levels(as.factor(df_param_all$neigh))), 
         key="species", value="abundance")   

community_id_df_projectiongraph  <- community_id_df %>% 
  mutate(family = case_when(!family %in% fam.to.keep ~ "rare",
                            T ~ family)) %>% 
  dplyr::filter(genus =="Lawrencella"|family %in% levels(as.factor(df_param_all$neigh))) %>%
  mutate(family = case_when(genus =="Lawrencella" ~"conspecific",
                            T~family)) %>%
  aggregate(count ~ id.plot + collector +  family + year + scale.width, sum) %>%
  mutate(count=count*scale.width.local.pred ) %>% 
  rename("species" = family,
        "abundance"= count) %>%
  dplyr::select(year, species, abundance)
  
community_id_df_projectiongraph.1 <- community_id_df_projectiongraph %>%
  mutate(function.name ="1.Traditional")
community_id_df_projectiongraph.2 <- community_id_df_projectiongraph %>%
  mutate(function.name ="2.Linear") %>%
  bind_rows(community_id_df_projectiongraph.1)
community_id_df_projectiongraph.3 <- community_id_df_projectiongraph %>%
  mutate(function.name ="3.Exp") %>%
  bind_rows(community_id_df_projectiongraph.2)
community_id_df_projectiongraph.4 <- community_id_df_projectiongraph %>%
  mutate(function.name ="4.Sigmoid")%>%
  bind_rows(community_id_df_projectiongraph.3)

allyear_communityprojection <- bind_rows(df_projection,
                                         community_id_df_projectiongraph.4 )


dummy <- data.frame(uplimit=c(60,60,60,60),
                    ylab=c(65,65,65,65),
                    significance=c("","","",""),
                    function.name = c("1.Traditional","2.Linear","3.Exp","4.Sigmoid"),
                    stringsAsFactors=FALSE)

family.neigh <-levels(as.factor(allyear_communityprojection$species))
family.neigh <-  family.neigh[!family.neigh =="conspecific"]

df_projection[which(df_projection$species=="conspecific"),] %>%
  aggregate(abundance ~ function.name, min)

plot_projection_list <- list()
plot_density_projection_list <- list()
library(ggridges)
for( fi in 1:4 ){ #"3.Exp"
  i <- c("1.Traditional","2.Linear","3.Exp","4.Sigmoid")[fi]
  title.vec <- c("a. Traditional constant",
                 "b. Linear","c. Exponential",
                 "d. Sigmoid")
  
  df <- allyear_communityprojection %>%
    dplyr::filter(function.name==i) %>%
    mutate(function.name=i) %>%
    mutate(species = factor(species, 
                            levels = c(family.neigh,"conspecific"))) %>%
    mutate(past = case_when( year < "2024" ~ "observed",
                             year > "2023" ~ "predicted"))
           

  plot_projection_list[[fi]] <- df  %>%
    ggplot() +
    annotate("rect", xmin=2010,xmax=2023,
            ymin=0,ymax=80,
             fill="lightgrey",alpha=0.3) + 
    annotate("text", x=2020, y=dummy[which(dummy$function.name==i),"ylab"],
               label=dummy[which(dummy$function.name==i),"significance"],
               color="black",size=16) + 
    stat_summary(data=df[which(df$species=="conspecific"),],
                 aes(y=abundance, x=year),
                 color="black",
                 alpha=0.9,size=1.5,
                 fun.y = mean,
                 fun.ymin = function(x) mean(x) -sd(x), 
                 fun.ymax = function(x) mean(x) +sd(x), 
                 geom = "pointrange") +
    stat_summary(aes(y=abundance, x=year,
                     group=species,color=species,
                     size =species, alpha=species),
                 fun.y = mean,
                 geom = "point") +
    stat_summary(aes(y=abundance, x=year,
                     group=species,color=species,
                     size =species, alpha=species),
                 fun.y = mean,
                 geom = "line") +
   scale_color_manual("",values=rev(cbp2),
                      labels=c("Araliaceae","Asteraceae","Crassulaceae",
                               "Goddeniaceae","Montiaceae","Poaceae","Rare families",

                               "LARO")) + 
   #scale_color_manual("",values=rev(c("black",rep("grey",
   #                                               times=length(family.neigh))))) +#values=cbp2) + 
   scale_size_manual("",values=rev(c(1.5, rep(1.2,length(family.neigh)))),
                     labels=c("Araliaceae","Asteraceae","Crassulaceae",
                              "Goddeniaceae","Montiaceae","Poaceae","Rare families",
                              "LARO")) + 
   scale_alpha_manual("",values=rev(c(0.9, rep(0.6,length(family.neigh)))),
                      labels=c("Araliaceae","Asteraceae","Crassulaceae",
                               "Goddeniaceae","Montiaceae","Poaceae","Rare families",
                               "LARO")) + 
    theme_bw() +
    labs(title = title.vec[fi],
         y= "",
         x="") +
    rremove("ylab") + rremove("xlab") + 
    guides(color= guide_legend("",position="bottom",
                               direction="horizontal",
                          byrow = TRUE,
                          nrow = 2),
           size= guide_legend(override.aes = list(size = 6)),
           linesize= guide_legend(override.aes = list(size = 2)),
           alpha= guide_legend(override.aes = list(alpha = 1))) +
    coord_cartesian( xlim = NULL, ylim = c(-1,40)) + #ylim =c(-5,80)) + #
    scale_x_continuous(expand= c(0,0),minor_breaks = NULL) +
    #scale_y_log10() +
    #scale_y_continuous(expand= c(0,0),
     #                  minor_breaks = NULL,
     #                  breaks=c(0,20,40,60,80),
     #                  labels=c(0,20,40,60,80)) +
    theme_bw() +
    theme(plot.title = element_text(color="#000000",
                                    vjust = 0,size=24),
          strip.background = element_blank(),
          strip.placement = "right",
          strip.text =element_text(color="black",
                                   size=22),
          legend.key= element_rect(fill = "white"),
          legend.text=element_text(size=18),
          legend.title=element_blank(),
          axis.text.x = element_text(color="black",size=16),
          axis.text.y = element_text(color="black",size=16),
          legend.key.size = unit(1, 'cm'),
          plot.margin = unit(c(0,-2,0,0), 'lines')) 
  plot_density_projection_list[[fi]]  <- ggplot() +
    geom_density_ridges2(data=df[which(df$species=="conspecific" &
                                         df$past =="observed"),],
                         aes(x=abundance,y=past, fill=past),
                         jittered_points = TRUE, 
            scale=1, alpha=0.8,
            point_size = 0.1, point_alpha = 0.05,
            position = position_raincloud(adjust_vlines = TRUE,
                                          height = 0.15,
                                          ygap = 0),
            quantile_lines = TRUE,
            quantiles = c( 0.5),
            vline_color = c("red"),vline_width=3)  +
    geom_density_ridges2(data=df[which(df$species=="conspecific" &
                                         df$past =="predicted"),],
                         aes(x=abundance,y=past, fill=past),
                         jittered_points = TRUE, 
                         scale=4,alpha=0.8,
                         point_size = 0.1, point_alpha = 0.05,
                         position = position_raincloud(adjust_vlines = TRUE,
                                                       height = 0.15,
                                                       ygap = 0),
                         quantile_lines = TRUE,
                         quantiles = c( 0.5),
                         vline_color = c("red"),vline_width=3)  +
    scale_fill_manual(values=c("lightgrey","black")) +
    coord_flip(xlim = c(-1,40), ylim = c(1,2), expand = c(0.01)) +
    guides(fill= guide_legend("",position="bottom",
                               direction="horizontal",
                               byrow = TRUE,
                               nrow = 2,
                              override.aes = list(quantile_lines = F))) + 
    theme_few() + 
    theme(plot.title = element_blank(), 
          legend.key= element_rect(fill = "white"),
          legend.text=element_text(size=18),
          panel.border = element_blank(),
          legend.title=element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          axis.title = element_blank(),
          legend.key.size = unit(1, 'cm'),
          panel.background = element_rect(fill='transparent'),
          plot.background = element_rect(fill='transparent', color=NA), #transparent plot bg
          panel.grid.major = element_blank(), #remove major gridlines
          panel.grid.minor = element_blank(), #remove minor gridlines
          plot.margin = unit(c(0,0,0,0), 'lines'))
  plot_density_projection_list[[fi]] 
}
plot_list <- list()
for(i in 1:4){
  plot_list[[i]] <- ggpubr::ggarrange( plot_projection_list[[i]] + theme(legend.position = "none",),
                                        plot_density_projection_list[[i]] +theme(legend.position = "none"),
                                        ncol=2,widths=c(1,0.5),
                                        align = c("h"))
}

plot_projection <- ggpubr::ggarrange(plotlist=plot_list,
                   ncol=1, nrow=4,heights = c(1,1,1,1),
                   common.legend = T,
                   legend="bottom",
                   align = c("hv"))
plot_projection
library(grid)
library(ggridges)
library(ggpubr)

plot_projection.2  <- annotate_figure(plot_projection  , 
                              left = textGrob("Mean stem density in 15 x 15cm", 
                                             rot = 90, vjust = 0.5,
                                              gp = gpar(fontsize=20,cex = 1.3)))+

  theme(plot.margin = unit(c(0,0,0,2), 'lines'))

plot_projection.2

legend.plot <- as_ggplot(ggpubr::get_legend(plot_projection_list[[4]]))
legend.density.plot <- as_ggplot(ggpubr::get_legend(plot_density_projection_list[[1]]))


plot_projection_legend <- ggarrange(plot_projection.2,
                                    heights=c(1,0.1),
                                   ggarrange(legend.plot,legend.density.plot,nrow=1,widths=c(1,0.3)),
                                   ncol=1) 
plot_projection_legend <-  annotate_figure(plot_projection_legend , 
                 bottom= textGrob("0.5",
                                  vjust = -6,hjust=-9.9,
                                  gp = gpar(fontsize=14,
                                            cex = 1.3,
                                            col="red"))) +
 theme(plot.margin = unit(c(0,0,-1,-1), 'lines'))
plot_projection_legend
ggsave(plot_projection_alpha , 
       file = "figures/NatData_plot_projection.pdf")


plot_projection

ggsave(plot_projection , 
       file = "figures/NatData_plot_poster.pdf")

#----4.3.Supp Graphs----
#----4.3.0.graph of Lambda----
grouping = "family"
df_lambda <- NULL
for(Code.focal in "LARO" ){ #focal.levels
  for (function.int in c(1:4)){
    df_alpha <- data.frame()
    function.vec <- c(0,0,0,0)
    function.vec[function.int] <- 1
    alpha.function <- paste0("function_",which(function.vec==1))
    
    load(paste0("results/stan/Parameters_",grouping,"_",Code.focal,"_",alpha.function,".RData"))
    parameters <- get(paste0("Parameters_",Code.focal,"_",alpha.function,"_",grouping))
    df_lambda <- bind_rows(df_lambda,
                           data.frame(function.int=alpha.function,
                                      lambda=parameters$lambda))
    
  }
}
str(df_lambda)
library(ggridges)
ggsave(df_lambda %>%
         mutate(function.name = case_when(function.int=="function_1" ~"1.Traditional",
                                          function.int=="function_2" ~"2.Linear",
                                          function.int=="function_3" ~"3.Exp",
                                          function.int=="function_4" ~"4.Sigmoid")) %>%
         ggplot(aes(x=lambda,y=function.name))  +
    geom_density_ridges(scale=1,quantile_lines = TRUE) +
    labs(y="function",x="intrinsic growth rate") +
  theme_bw(),
  file="figures/NatCom_lambda.pdf")

#----4.3.1.graph of LARO's population density----
str(community_id_df_LARO)
ggsave(community_id_df_LARO %>%
         filter(count>0) %>%
         ggplot() +
         geom_histogram(aes(x=count),fill="pink",binwidth=2)+
         labs(x=expression(number~of~individuals~per~15~"cm"^2))+
         coord_cartesian(xlim=c(0,45)) + 
         theme_clean() +
         theme(axis.text = element_text(size=24),
               axis.title = element_text(size=24),
               axis.ticks = element_line(size=3)),
       file="figures/NatCom_LARODensity.pdf")

#----4.3.2.graph of neighbours population density----
library(ggridges)
ggsave(community_id_df_fam %>%
         gather(any_of(c(fam.to.keep,"rare")),
                key="neigh",value="abundance") %>%
         filter(!is.na(abundance)) %>%
         ggplot(aes(x=abundance,y=neigh)) +
         geom_density_ridges(stat = "binline", 
                             bins = 1000, 
                             scale = 0.95, draw_baseline = FALSE) +
         labs(y="",x="Density in a 15 by 15 cm squarre") +
         coord_cartesian(xlim = c(0,50)) +
         theme_bw()+
         theme(axis.text = element_text(size=14),
               axis.title = element_text(size=16)),
       file="figures/NatCom_NeighbourhoodDensity.pdf")

#----4.3.3.graph of realised effect on lambda----
str(community_id_df_fam)

realised.effect.df <- community_id_df %>% 
  mutate(family = case_when(!family %in% fam.to.keep ~ "rare",
                            T ~ family)) %>%
  dplyr::filter(genus =="Lawrencella"|family %in% levels(as.factor(df_param_all$neigh))) %>%
  mutate(family = case_when(genus =="Lawrencella" ~"conspecific",
                            T~family)) %>%
  aggregate(count ~ id.plot + collector +  family + year + scale.width, sum) %>%
  mutate(count=count*scale.width.local.pred ) %>% 
  rename("neigh" = family,
         "abundance"= count) %>%
  dplyr::select(year, neigh, abundance) %>%
  full_join(df_param_all, relationship="many-to-many")

levels(as.factor(realised.effect.df$neigh))

str(realised.effect.df)
view(realised.effect.df)
for (n in 1:nrow(realised.effect.df)){
  print(n)
  if(realised.effect.df$function.int[n]==1){
    realised.effect.df$alpha_value[n] <- realised.effect.df$alpha_init[n] 
    
  }
  if(realised.effect.df$function.int[n]==2){
    realised.effect.df$alpha_value[n] <- alpha_function2(realised.effect.df$alpha_init[n],
                                                         realised.effect.df$alpha_slope[n],
                                                         realised.effect.df$abundance[n],
                                                         realised.effect.df$Nmax[n])
    
  }
  if(realised.effect.df$function.int[n]==3){
    realised.effect.df$alpha_value[n]<- alpha_function3(realised.effect.df$alpha_init[n],
                                                        realised.effect.df$alpha_slope[n],
                                                        realised.effect.df$alpha_c[n],
                                                        realised.effect.df$abundance[n],
                                                        realised.effect.df$Nmax[n])
    
  }
  if(realised.effect.df$function.int[n]==4){
    realised.effect.df$alpha_value[n] <- alpha_function4(realised.effect.df$alpha_init[n],
                                                         realised.effect.df$alpha_slope[n],
                                                         realised.effect.df$alpha_c[n],
                                                         realised.effect.df$abundance[n],
                                                         realised.effect.df$Nmax[n])
  }
}

ggsave(realised.effect.df %>%
  mutate(neigh = factor(neigh, 
                        levels = c(family.neigh,"conspecific"))) %>%
  mutate(function.name = case_when(function.int==1 ~"1.Traditional",
                                 function.int==2 ~"2.Linear",
                                 function.int==3 ~"3.Exp",
                                 function.int==4 ~"4.Sigmoid")) %>%
  mutate(realised.effect = exp(alpha_value*abundance)) %>%
  ggplot(aes(y=realised.effect ,x=neigh,
             fill=neigh)) +
  geom_boxplot() +
  geom_hline(yintercept=1) + 
  scale_fill_manual("",values=rev(cbp2),
                    labels=c("Araliaceae","Asteraceae","Crassulaceae",
                             "Goddeniaceae","Montiaceae","Poaceae","Rare families","LARO")) +
  facet_grid(.~function.name)  +
  coord_cartesian(ylim=c(0,3)) +
  guides(fill=guide_legend("",position="bottom",
                            direction="horizontal",
                            byrow = TRUE,
                            nrow = 2)) +
  labs(y="Realised effect on the intrinsic growth rate,
        weighted by the observed abundances in the last 10 years")+
  theme_bw()+  
  theme(legend.position="bottom",
        axis.text.x = element_blank(),
        strip.text =element_text(color="black",
                                 size=22),
        legend.key= element_rect(fill = "white"),
        legend.text=element_text(size=18),
        legend.title=element_blank(),
        axis.title = element_text(color="black",size=20),
        axis.text.y = element_text(color="black",size=16),
        legend.key.size = unit(1, 'cm'),
        plot.margin = unit(c(1,1,1,1), 'lines')),
  file="figures/Nat.com.realised.effect.pdf")
  
#----4.3.4.Table of projection summary----
summary.proj <- df_projection[which(df_projection$species=="conspecific"),] %>%
  bind_rows(community_id_df_LARO %>% mutate(function.name="observed") %>%
              rename("abundance"="count")) %>%
  group_by(function.name) %>% 
  mutate(ab.mean = mean(abundance, na.rm = T),
         ab.max = max(abundance, na.rm = T),
         ab.min = min(abundance, na.rm = T),
         ab.median =median(abundance,na.rm=T),
         ab.sd =sd(abundance,na.rm=T),
         dfreedom= length(abundance)-1) %>%
  dplyr::select(function.name,ab.min,ab.max,ab.mean,ab.sd,ab.median,dfreedom) %>%
  unique() %>%
  ungroup() 
view(summary.proj)
str(community_id_df_LARO )

write.csv(summary.proj, file="results/summary.proj.csv")

# Graph of density distribution 
library(ggridges)
ggsave(
  ggplot() +
    geom_density(data=df_projection[which(df_projection$species=="conspecific"),],
                 aes(x=abundance,y=..scaled.., 
                     fill=function.name),
                 alpha=0.5,size=1) +
    facet_grid(function.name~.) +
    geom_density(data=community_id_df_LARO,
                 aes(x=count,y=..scaled..),
                 color="black",size =1.5,
                 alpha=0)+
    labs(x="number of individuals per 15cm",
         y="scaled frequency") +
    guides(fill="none") +
    coord_cartesian(xlim=c(0,50)) + 
    theme_clean() +
    theme(axis.text = element_text(size=14),
          axis.title = element_text(size=16),
          strip.text =element_text(color="black",
                                   size=22)) ,
  file="figures/density.proj.fam.pdf")


#---- 4.4. Compare model for each focal  ----
# reference : Vehtari, A., Gelman, A., and Gabry, J. (2017). Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC. Statistics and Computing. 27(5), 1413–1432. :10.1007/s11222-016-9696-4. online, arXiv preprint arXiv:1507.04544.
model.loo <- list()

for(Code.focal in "LARO"){
  for (function.int in 1:4){ # c(1:4)
    if(ModelCheck[which(ModelCheck$function.int==function.int &
                       ModelCheck$focal==Code.focal),"Rhat"] < 0 | 
      ModelCheck[which(ModelCheck$function.int==function.int &
                         ModelCheck$focal==Code.focal),"Neff"] < 100) next
      
      print(paste(Code.focal,", function",function.int))
      
      load(paste0("results/stan/FinalFit_family_",Code.focal,"_function_",function.int,".rds"))
      # Extract pointwise log-likelihood
      # using merge_chains=FALSE returns an array, which is easier to 
      # use with relative_eff()
      log_lik <- loo::extract_log_lik(FinalFit, 
                                      parameter_name = "F_sim", 
                                      merge_chains = F)
      #as of loo v2.0.0 we can optionally provide relative effective sample sizes
      # when calling loo, which allows for better estimates of the PSIS effective
      # sample sizes and Monte Carlo error
      r_eff <- loo::relative_eff(log_lik, cores = 3) 
      # preferably use more than 2 cores (as many cores as possible)
      # will use value of 'mc.cores' option if cores is not specified
      model.loo[[paste0(Code.focal,"_function_",function.int)]] <-  loo::loo(log_lik,
                                                                            threshold=0.7,
                                                                            r_eff = r_eff, 
                                                                            cores = 3)
      remove(FinalFit)
  }
}

Se_loo_model <- NULL
for(Code.focal in "LARO"){
  Se_loo_model <- loo::loo_compare(model.loo[[paste0(Code.focal,"_function_1")]], 
                        model.loo[[paste0(Code.focal,"_","function_2")]],
                        model.loo[[paste0(Code.focal,"_","function_3")]],
                        model.loo[[paste0(Code.focal,"_","function_4")]])
  

  Se_loo_model <-as.data.frame(print(Se_loo_model, simplify = FALSE, digits = 3))
  
  }
Se_loo_model<- Se_loo_model %>%
  mutate(across(where(is.numeric), \(x) tibble::num(x, digits = 3))) %>%
  rownames_to_column(var="model") #%>%
  #separate(col=function.int, 
  #         into=c("model","function.int"),sep="l") %>%
  #dplyr::subset(-model)
view(Se_loo_model)
write_csv(as.data.frame(Se_loo_model),
          paste0("results/LARO_Se_loo_model.csv"))

#---- 4.5. Figure of fecundity estimation distribution ----
source("code/6.1.Stan_modelcheck_rem.R")
function.names <- c("1.Traditional","2.Linear","3.Exp","4.Sigmoid")
  pdf(paste0("figures/NatData_PostFecundity_distribution.pdf")) 
  par(oma=c(0,0,0,0), mar=c(2.25,2.5,2,1))
  layout(mat = matrix(
    1:4,
    byrow=TRUE,
    nrow = 1,
    ncol = 4
  ),
  heights = rep(1,8),
  widths = rep(4,8)
  )
  
  for(Code.focal in "LARO"){ #,"j"
    for (function.int in c(1:4)){ # c(1:4)
     #col2 <- c("black","#CC79A7","#E69F00","#009E73") # function 
      col2 <- wes_palette("FantasticFox1", n = 5)
      #col1 <- c("blue","red") # focal species
      # loo values
      value.se <- as.numeric(Se_loo_model$looic[which(Se_loo_model$model== paste0("model",function.int))])
      # fecundity obs
      load(paste0("results/stan/Fsim_family_",Code.focal,"_function_",function.int,".RData"))
      # fecundity generated
      Fec_df <- data.frame(Obs = 1:get(paste0("Fsim_",Code.focal,"_function_",function.int,"_",grouping))[["DataVec"]]$N,
                           Fec =get(paste0("Fsim_",Code.focal,"_function_",function.int,"_",grouping))[["DataVec"]]$Fecundity)
      
      
      load(paste0("results/stan/FinalFit_family_",Code.focal,"_function_",function.int,".rds"))
      FinalPosteriors <- rstan::extract(FinalFit)
      stan_post_pred_check_all(FinalPosteriors,"F_hat",
                               Fec_df$Fec,
                               function.names[function.int],
                               "black",
                               col2[1+function.int],value.se
                               ) 
      
      
    }
  }
  dev.off()
